turn observation action buttons into shared component
